Warren and Nostown Remap! (#402)
* Warren + Nostown Remap

Added, and upgraded some structures in the underground city, as well as expanding upon the maze

* Update SanFrancisco.dmm

* slight edits

playtested it again, deleted some floating cameras and turned off the music in the bar so they can play whatever music they want without overlapping the theme

* Warrens update

added two bathrooms to 2 of the houses, plus some decor.
plus one closet in the house furthest in the back

* slight edits

deleted a thing i forgot to take out in an out of bounds area and adjusted the zones

* attempted merge conflict fix

added the endron map and some other prs
fixed the sprinkler system
redecorated some rooms to make it look a little more lived in
realized putting stuff ontop of crates makes it so that said stuff goes INSIDE the crates... Went back and edited some of that.

* Update SanFrancisco.dmm

* Update SanFrancisco.dmm

Tried fixing the sprinklers
replaced some tables that fling you off of them whenever you try to jump at them
decorated again

* Update SanFrancisco.dmm

i hate sprinklers

* Update SanFrancisco.dmm

* Update SanFrancisco.dmm

Took away music in the nosbar so the music won't overlap

* zones

added nosferatu warrens, and nosferatu bar as zones to try and fix the sprinkler music issue

* Update SanFrancisco.dmm

redecorated some of the interiors, fixed some wonky lighting clipping into the walls.

* Update SanFrancisco.dmm

* tried fixing the sofas and the coat hangers

* Update SanFrancisco.dmm

* fix attempt

another attempt to fix

* Playtest fixes

Checked to see if the gates worked, (they do)
Checked what was up with the sisters of  cacophany table, everything on there worked as normal
added a chair to the primo nos office
re-edited some grates people werent supposed to walk over

* Fixes

- changed the zones, People can no longer link their radios while inside the store, and apartments.
- Added a new office zone, lets people connect to the radio in the security room + primo office in the warrens
- Added some railings around the basketball court
- Removed meth lab and blood extractor machine
- Added a trophy for shits and gigs and for whoever wins a match while playing

* Update SanFrancisco.dmm

* Update SanFrancisco.dmm

* Update SanFrancisco.dmm

* Update SanFrancisco.dmm

* Update SanFrancisco.dmm

* Update SanFrancisco.dmm

* Update SanFrancisco.dmm

* Update SanFrancisco.dmm

* deleted floating clocks

* floating lightbulb and random deer evicted

* ambient noise fix + slight edits

---------

Co-authored-by: XeonMations <62395746+XeonMations@users.noreply.github.com>