TGUI runtime fixes (#734)
* Fix several SHOULD_NOT_SLEEP hits in SStgui.update_uis() (#75411)

There are several esoteric code paths that lead to `winexists` calls or
sleeps in `ui_data`. This proc is meant to be a command without
sleeping.

See #75232

![image](https://github.com/tgstation/tgstation/assets/8171642/8becd881-d7e8-4fe5-90af-2c2657934d07)

* Fixes some spurious runtimes in SStgui procs (#76251)

## About The Pull Request

src_object is NOT guaranteed to be passed in. HHHHHHHHHHH
Closes #76249

* Fixes tgui_open_uis tracking (#77101)

Fixes https://github.com/tgstation/tgstation/issues/77097

This breaks uis for everyone with fancy-tgui off and in cases where the
window is closed by external action (byond closing it) which is pretty
bad.

* Update interface.dm

---------

Co-authored-by: Jordan Dominion <Cyberboss@users.noreply.github.com>
Co-authored-by: LemonInTheDark <58055496+LemonInTheDark@users.noreply.github.com>
Co-authored-by: AnturK <AnturK@users.noreply.github.com>
Co-authored-by: chazzyjazzy <33268885+chazzyjazzy@users.noreply.github.com>